구조 재수정
