+++ app/src/main/java/com/example/weatherapp/main/view/MainActivity.kt
@@ -8,6 +8,5 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
     }
 }
