+++ app/src/main/java/com/example/weatherapp/repository/WeatherRepository.kt
@@ -1,9 +1,5 @@
 package com.example.weatherapp.repository
 
-import android.content.ContentValues.TAG
-import android.util.Log
-import retrofit2.Call
-import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import java.time.LocalDate
@@ -17,34 +13,9 @@ class WeatherRepository {
         .build()
 
     private val api = retrofit.create(WeatherAPI::class.java)
-    fun getWeather(city: Int, callback: (weathers: List<WeatherResult>) -> Unit) {
-        for (i in 0 until 6) {
-            val nextDay = time.plusDays(i.toLong())
-            val callGetWeathers = api.getWeather(city, nextDay.year, nextDay.monthValue, nextDay.dayOfMonth)
-            callGetWeathers.enqueue( // enqueue 로 비동기 실행
-                object : retrofit2.Callback<List<WeatherResult>> {
-                    override fun onResponse(
-                        call: Call<List<WeatherResult>>,
-                        response: Response<List<WeatherResult>>
-                    ) {
-                        response.body()?.let {
-                            callback.invoke(it.slice(0..0))
-                        }
-                    }
-
-                    override fun onFailure(call: Call<List<WeatherResult>>, t: Throwable) {
-                        Log.d(TAG, "onFailure: " + t.message)
-                    }
-
-
-                }
-            )
-        }
-    }
 
     suspend fun getWeatherAsync(city: Int, year: Int, month: Int, day: Int): List<WeatherResult> =
         api.getWeatherAsync(city, year, month, day).slice(0..0)
-
 }
